deps/v8/test/inspector/debugger/wasm-set-breakpoint-breaks-on-first-breakable-location-expected.txt

Tests if breakpoint set is first breakable location

Running test: test
Running test function...
Set breakpoint outside of any function: (0, 0).
Setting breakpoint for id: 4 at 0, 0.
{
    error : {
        code : -32000
        message : Could not resolve breakpoint
    }
    id : <messageId>
}
Set breakpoint could not resolve break location.
Set breakpoint at a breakable location: (0, 40).
Setting breakpoint for id: 4 at 0, 40.
Location match for (0, 40).
Initial location is expected to be breakable: true.
Set breakpoint at non-breakable location: (0, 42).
Setting breakpoint for id: 4 at 0, 42.
Location match for (0, 43).
Initial location is expected to be breakable: false.
